requirements.txt
numpy
pillow
scikit-learn
tqdm
albumentations==1.3.0
jupyterlab
matplotlib
natsort
scikit-image>=0.16.1
tensorboardx
tensorboard
torchcontrib
tifffile
pygit2
open3d